+++ src/TuDa.CIMS.Web/Components/WorkingGroupPage/WorkingGroupHeader.razor.cs
@@ -1,8 +1,7 @@
 ﻿using Microsoft.AspNetCore.Components;
 using TuDa.CIMS.Shared.Entities;
 
-namespace TuDa.CIMS.Web.Components.WorkingGroup;
-
+namespace TuDa.CIMS.Web.Components.WorkingGroupPage;
 public partial class WorkingGroupHeader : ComponentBase
 {
     [Parameter] public required Professor Professor { get; set;}
