+++ src/TuDa.CIMS.Web/Components/WorkingGroupPage/WorkingGroupInvoiceButton.razor.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.AspNetCore.Components;
 
-namespace TuDa.CIMS.Web.Components.WorkingGroup;
-
+namespace TuDa.CIMS.Web.Components.WorkingGroupPage;
 public partial class WorkingGroupInvoiceButton : ComponentBase
 {
     [Parameter] public EventCallback Checkout { get; set; }
