+++ src/TuDa.CIMS.Web/Components/WorkingGroupPage/WorkingGroupPurchaseList.razor.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Components;
 using TuDa.CIMS.Shared.Entities;
 
-namespace TuDa.CIMS.Web.Components.WorkingGroup;
+namespace TuDa.CIMS.Web.Components.WorkingGroupPage;
 
 public partial class WorkingGroupPurchaseList : ComponentBase
 {
